refactor: introduce src/pdi-common-impls
