+++ src/Game.ts
@@ -1,7 +1,5 @@
 import * as pl from "@akashic/playlog";
 import { Trigger } from "@akashic/trigger";
-import { ExceptionFactory } from "./commons/ExceptionFactory";
-import { SurfaceAtlasSet } from "./commons/SurfaceAtlasSet";
 import { AssetManager } from "./domain/AssetManager";
 import { AudioSystemManager } from "./domain/AudioSystemManager";
 import { Camera } from "./domain/Camera";
@@ -15,9 +13,11 @@ import { OperationPluginManager } from "./domain/OperationPluginManager";
 import { PointEventResolver } from "./domain/PointEventResolver";
 import { RandomGenerator } from "./domain/RandomGenerator";
 import { Storage } from "./domain/Storage";
+import { SurfaceAtlasSet } from "./domain/SurfaceAtlasSet";
+import { SurfaceAtlasSetLike } from "./domain/SurfaceAtlasSetLike";
 import { XorshiftRandomGenerator } from "./domain/XorshiftRandomGenerator";
 import { GameHandlerSet } from "./GameHandlerSet";
-import { SurfaceAtlasSetLike } from "./interfaces/SurfaceAtlasSetLike";
+import { ExceptionFactory } from "./pdi-common-impls/ExceptionFactory";
 import { AssetLike } from "./pdi-types/AssetLike";
 import { CommonOffset, CommonSize } from "./pdi-types/commons";
 import { PlatformPointEvent, PlatformPointType } from "./pdi-types/PlatformPointEvent";
