+++ src/domain/LoadingScene.ts
@@ -1,5 +1,5 @@
 import { Trigger } from "@akashic/trigger";
-import { ExceptionFactory } from "../commons/ExceptionFactory";
+import { ExceptionFactory } from "../pdi-common-impls/ExceptionFactory";
 import { AssetLike } from "../pdi-types/AssetLike";
 import { Scene, SceneParameterObject } from "../Scene";
