+++ src/domain/SpriteFactory.ts
@@ -1,4 +1,4 @@
-import { ExceptionFactory } from "../commons/ExceptionFactory";
+import { ExceptionFactory } from "../pdi-common-impls/ExceptionFactory";
 import { Scene } from "../Scene";
 import { Camera } from "./Camera";
 import { E } from "./entities/E";
