+++ src/domain/TimerManager.ts
@@ -1,5 +1,5 @@
 import { Trigger } from "@akashic/trigger";
-import { ExceptionFactory } from "../commons/ExceptionFactory";
+import { ExceptionFactory } from "../pdi-common-impls/ExceptionFactory";
 import { Timer } from "./Timer";
 
 /**
