+++ src/domain/XorshiftRandomGenerator.ts
@@ -1,4 +1,4 @@
-import { ExceptionFactory } from "../commons/ExceptionFactory";
+import { ExceptionFactory } from "../pdi-common-impls/ExceptionFactory";
 import { RandomGenerator } from "./RandomGenerator";
 import { Xorshift, XorshiftSerialization } from "./Xorshift";
