+++ src/index.common.ts
@@ -3,13 +3,13 @@ export * from "@akashic/trigger";
 export * from "./Scene";
 export * from "./Game";
 
-export * from "./commons/ExceptionFactory";
-export * from "./commons/Glyph";
-export * from "./commons/ShaderProgram";
-export * from "./commons/SurfaceAtlasSet";
-export * from "./commons/SurfaceAtlas";
-export * from "./commons/SurfaceAtrasSlot";
-export * from "./commons/VideoSystem";
+export * from "./pdi-common-impls/ExceptionFactory";
+export * from "./pdi-common-impls/Glyph";
+export * from "./domain/ShaderProgram";
+export * from "./domain/SurfaceAtlasSet";
+export * from "./pdi-common-impls/SurfaceAtlas";
+export * from "./pdi-common-impls/SurfaceAtrasSlot";
+export * from "./domain/VideoSystem";
 
 export * from "./domain/entities/CacheableE";
 export * from "./domain/entities/E";
@@ -54,19 +54,19 @@ export * from "./domain/Util";
 export * from "./domain/Xorshift";
 export * from "./domain/XorshiftRandomGenerator";
 
-export * from "./implementations/Asset";
-export * from "./implementations/AudioAsset";
-export * from "./implementations/AudioPlayer";
-export * from "./implementations/AudioSystem";
-export * from "./implementations/GlyphFactory";
-export * from "./implementations/ImageAsset";
-export * from "./implementations/Renderer";
-export * from "./implementations/ResourceFactory";
-export * from "./implementations/ScriptAsset";
-export * from "./implementations/Surface";
-export * from "./implementations/TextAsset";
-export * from "./implementations/VideoAsset";
-export * from "./implementations/VideoPlayer";
+export * from "./pdi-common-impls/Asset";
+export * from "./pdi-common-impls/AudioAsset";
+export * from "./pdi-common-impls/AudioPlayer";
+export * from "./pdi-common-impls/AudioSystem";
+export * from "./pdi-common-impls/GlyphFactory";
+export * from "./pdi-common-impls/ImageAsset";
+export * from "./pdi-common-impls/Renderer";
+export * from "./pdi-common-impls/ResourceFactory";
+export * from "./pdi-common-impls/ScriptAsset";
+export * from "./pdi-common-impls/Surface";
+export * from "./pdi-common-impls/TextAsset";
+export * from "./pdi-common-impls/VideoAsset";
+export * from "./pdi-common-impls/VideoPlayer";
 
 export * from "./pdi-types/AssetLike";
 export * from "./pdi-types/AssetLoadFailureInfo";
@@ -83,7 +83,7 @@ export * from "./pdi-types/ScriptAssetRuntimeValue";
 export * from "./pdi-types/ScriptAssetLike";
 export * from "./pdi-types/ShaderProgramLike";
 export * from "./pdi-types/SurfaceAtlasLike";
-export * from "./interfaces/SurfaceAtlasSetLike";
+export * from "./domain/SurfaceAtlasSetLike";
 export * from "./pdi-types/SurfaceAtlasSlotLike";
 export * from "./pdi-types/SurfaceLike";
 export * from "./pdi-types/TextAssetLike";
@@ -94,7 +94,7 @@ export * from "./pdi-types/VideoSystemLike";
 export * from "./pdi-types/AudioAssetHint";
 export * from "./pdi-types/ImageAssetHint";
 export * from "./types/AssetConfiguration";
-export * from "./types/AssetLoadErrorType";
+export * from "./pdi-types/AssetLoadErrorType";
 export * from "./pdi-types/commons";
 export * from "./pdi-types/CompositeOperation";
 export * from "./pdi-types/CompositeOperationString";
