+++ src/pdi-common-impls/GlyphFactory.ts
