+++ src/pdi-common-impls/ScriptAsset.ts
