+++ src/pdi-common-impls/SurfaceAtrasSlot.ts
