+++ src/pdi-types/AssetLoadErrorType.ts
